# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zeitgeist",
    "version": "0.1.0"
  },
  "paths": {
    "/api/runs": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Start Run",
        "description": "202 rather than 201: the run is queued, and for a queued one there is\nnothing at its URL yet.",
        "operationId": "start_run_api_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Resume Run",
        "operationId": "resume_run_api_runs__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/stop": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Stop Run",
        "description": "Finish the current stage, write its checkpoint, then end. The run\nstays resumable, which is what the button promises.",
        "operationId": "stop_run_api_runs__run_id__stop_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunActionAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/abort": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Abort Run",
        "description": "End now. Ingest is the exception: `fetch_evidence` is one opaque\n`asyncio.run()` with no interior checkpoint, so an abort during it takes\neffect when the fetch returns \u2014 which is what the UI's \"aborting\u2026\" state\nis for.",
        "operationId": "abort_run_api_runs__run_id__abort_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunActionAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "tags": [
          "control"
        ],
        "summary": "Stream Events",
        "description": "Log lines and progress ticks, until the run ends.\n\nPolls the run's buffer rather than being pushed to. A quarter-second of\nlatency on a log line is invisible, and polling keeps the logging handler\nfree of any reference to the event loop \u2014 which is what lets the same\nhandler work under `TestClient` and under a real server alike.\n\nA tick carries no payload. The observer has already written\n`run_records`, `run_stages` and `run_topics` from the worker thread, so\nthe client refetches rather than being handed the same data twice.",
        "operationId": "stream_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/renders/{render_id}": {
      "get": {
        "tags": [
          "renders"
        ],
        "summary": "Read Render",
        "operationId": "read_render_api_renders__render_id__get",
        "parameters": [
          {
            "name": "render_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Render Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "renders"
        ],
        "summary": "Remove Render",
        "description": "204 and no body. The row goes first, then both files.\n\nA render whose PNG is already missing still deletes cleanly: the row is\nwhat makes it exist, and the end state is the same either way.\n\nThe meme count every screen shows is `COUNT(*)` over `renders` at query\ntime, so nothing else has to be told this happened \u2014 which is the\nreason that count is not a column.",
        "operationId": "remove_render_api_renders__render_id__delete",
        "parameters": [
          {
            "name": "render_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Render Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoOrigin": {
        "properties": {
          "provenance": {
            "type": "string",
            "const": "auto",
            "title": "Provenance",
            "default": "auto"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rationale"
        ],
        "title": "AutoOrigin",
        "description": "A render the model briefed: it chose the template and said why."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualOrigin": {
        "properties": {
          "provenance": {
            "type": "string",
            "const": "manual",
            "title": "Provenance",
            "default": "manual"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ManualOrigin",
        "description": "A render written by hand. No model call, so no choice to explain."
      },
...
      "QueuedRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "position"
        ],
        "title": "QueuedRun",
        "description": "`position` 0 means executing now; higher means waiting behind that\nmany runs, which is what the \"queues it behind\" notice reads."
      },
...
      "RenderRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "topic_id": {
            "type": "string",
            "title": "Topic Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "caption_slots": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Caption Slots"
          },
          "origin": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AutoOrigin"
              },
              {
                "$ref": "#/components/schemas/ManualOrigin"
              }
            ],
            "title": "Origin",
            "discriminator": {
              "propertyName": "provenance",
              "mapping": {
                "auto": "#/components/schemas/AutoOrigin",
                "manual": "#/components/schemas/ManualOrigin"
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "generating",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "topic_id",
          "template_id",
          "caption_slots",
          "origin",
          "status",
          "error",
          "created_at"
        ],
        "title": "RenderRecord",
        "description": "One rendered meme. The database is authoritative for whether it exists;\nthe PNG and its thumbnail live at `output/<run_id>/renders/<id>.png`.\n\nWhile `status` is `\"generating\"`, `caption_slots` and `origin`'s\n`rationale` (on `AutoOrigin`) are not yet written \u2014 they hold\nplaceholder values until the job finishes drawing this render \u2014 and\nmust not be read as its real captions or the model's real reasoning.\n\n`template_id` is None where no template has been chosen: a `generating`\nrow whose request left the choice to the model, and a `failed` row\nwhose brief failed before the model chose. That is a real state rather\nthan a placeholder \u2014 there is no template to name \u2014 so it is `| None`\nrather than an empty string standing in for one."
      },
...
      "ResumeBody": {
        "properties": {
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stage"
              },
              {
                "type": "null"
              }
            ]
          },
          "template_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ResumeBody",
        "description": "`stage` omitted means the computed resume point \u2014 the button posts no\nstage. `template_ids` narrows the library for this resume only, which is\nthe tuning loop: edit a manifest, re-render the same frozen topics."
      },
      "RunActionAck": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "requested": {
            "type": "string",
            "enum": [
              "stop",
              "abort"
            ],
            "title": "Requested"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "requested"
        ],
        "title": "RunActionAck",
        "description": "Acknowledges `POST /{run_id}/stop` and `/abort`.\n\nBoth endpoints returned a bare `dict[str, str]`, unlike every other\nendpoint in the project, which declares a `response_model`. Phase 5\ngenerates a TypeScript client from this OpenAPI schema, and two\nundeclared response bodies out of fifteen endpoints is a client-visible\nhole this closes."
      },
...
      "RunPage": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runs",
          "next_cursor"
        ],
        "title": "RunPage",
        "description": "One page of runs. `next_cursor` is None on the last page."
      },
      "RunRecordRow": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "ok",
              "failed",
              "aborted",
              "interrupted"
            ],
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "attempt_started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Attempt Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "config": {
            "$ref": "#/components/schemas/RunConfig"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunError"
              },
              {
                "type": "null"
              }
            ]
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "trends_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trends Found"
          },
          "topics_kept": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topics Kept"
          },
          "phrases_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phrases Found"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "started_at",
          "attempt_started_at",
          "finished_at",
          "config",
          "error",
          "item_count",
          "trends_found",
          "topics_kept",
          "phrases_found"
        ],
        "title": "RunRecordRow",
        "description": "A row of `run_records`, read back.\n\nThe counts are `| None` because they are written when the run finishes;\na run still in flight has not counted anything yet."
      },
      "RunSummary": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunRecordRow"
          },
          "topic_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topic Labels"
          },
          "render_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Render Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run",
          "topic_labels",
          "render_ids"
        ],
        "title": "RunSummary",
        "description": "One row of the Runs list.\n\n`topic_labels` and `render_ids` are joins the row needs and the run\nrecord does not carry: column two is the titles, column four the\nthumbnails."
      },
...
      "Stage": {
        "type": "string",
        "enum": [
          "ingest",
          "analyse",
          "evaluate",
          "generate"
        ],
        "title": "Stage"
      },
...
      "StartRunBody": {
        "properties": {
          "template_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Ids"
          },
          "overrides": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Overrides",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StartRunBody",
        "description": "What `POST /api/runs` accepts.\n\n`overrides` is `str`-valued because the values come from a form and are\nvalidated by `Settings` when the worker freezes the run's config \u2014 the\none place that knows each field's real type."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}